Restructured UserDirectories folder and Prep for GenericLdapUserDirectory
