+++ InedoCore/InedoExtension/UserDirectories/LdapHelperV4.cs
@@ -9,9 +9,9 @@ namespace Inedo.Extensions.UserDirectories
     ****************************************************************************************************/
     internal static class LdapHelperV4
     {
-        private static readonly LazyRegex LdapEscapeRegex = new LazyRegex(@"[,\\#+<>;""=]", RegexOptions.Compiled);
-        private static readonly LazyRegex LdapUnescapeRegex = new LazyRegex(@"\\([,\\#+<>;""=])", RegexOptions.Compiled);
-        private static readonly LazyRegex LdapSplitRegex = new LazyRegex(@"(?<!\\),", RegexOptions.Compiled);
+        private static readonly LazyRegex LdapEscapeRegex = new(@"[,\\#+<>;""=]", RegexOptions.Compiled);
+        private static readonly LazyRegex LdapUnescapeRegex = new(@"\\([,\\#+<>;""=])", RegexOptions.Compiled);
+        private static readonly LazyRegex LdapSplitRegex = new(@"(?<!\\),", RegexOptions.Compiled);
 
         public static string Escape(string s)
         {
